Mix between new version and old Nirjal version
Had to do this mix between new and old version of the DarwinDrop contract to resolve mismatchings with the frontend. Kieran confirmed I could remove the param "payedWithDarwin" (that was one of the params causing mismatching), which we will reintroduce in the future and reflect its utilization on frontend. Other removed params, like tokensPerUser, were actually useless.